Break up map component into subcomponents
