+++ src/types/maps.ts
@@ -102,3 +102,17 @@ export type SubmapData = {
   vmax: number | undefined;
   cmap: string | undefined;
 };
+
+export type BoxWithPositionalData = Box & {
+  top: number;
+  left: number;
+  width: number;
+  height: number;
+};
+
+export type NewBoxData = Omit<Box, 'id' | 'name' | 'description'> & {
+  top: number;
+  left: number;
+  width: number;
+  height: number;
+};
